Refactor mortality utilities and add isotonic calibration
